Начало, на основе прошивки 81.20.75R от aura(nevis)

+++ linux-2.6.x/net/atm/addr.c
@@ -114,7 +114,8 @@ int atm_del_addr(struct atm_dev *dev,struct sockaddr_atmsvc *addr)
 }
 
 
-int atm_get_addr(struct atm_dev *dev,struct sockaddr_atmsvc __user *buf,int size)
+int atm_get_addr(struct atm_dev *dev, struct sockaddr_atmsvc __user * buf,
+		 size_t size)
 {
 	unsigned long flags;
 	struct atm_dev_addr *walk;
